A-10 DataLoader: Fix duplicates, Add prefix, pad waypoints
To avoid issues with existing waypoints (e.g. Al Dhafra) which case them to be
added as MSNXXXX, we prefix all with "M "

Additionally to avoid issues with autocomplete, we suffix all waypoints with
spaces such that waypoints entered that contain previously entered waypoints
can't be autocompleted:

  E.g:

    Racetrack
    Race

This causes 2 issues:

  Waypoint Entry:
    Race autocompletes to Racetrack, causing it to be called MSNXXXX

  Flight Plan:
    Race again autocompletes to Racetrack, causing an incorrect point to be
    added to the FlightPlan which isn't what we would want.

Spaces seem the simplest workaround without having to educate.