Fix #15: Add LSS and set default
LSS is now defaulted to the element lead and lead to wing.

Of note if we have a 5 ship, it will be a lead element of 2 ships, and
secondary element of 3 ships rather than a 3+2 and the MDC author will need to
take care of the LSS themselves.

Due to the pain involved with deleting a middle row (both in the ID, OID,
SQUAWK etc) i've decided to only allow deletion from the end, to encourage
someone to rename a pilot slot than have to create more pain for themselves

This also fixes an issue that was created when moving from named arguments to
object based arguments - where the previously saved values weren't being
correctly re-populated if they were manually adjusted