Fix #21: Add freeform text field for Loadout Information
For some ordnance and in some missions, we are adjusting the default DSMS
profiles with custom parameters, or program new profiles

There are numerous options to make, so it is not really sensible to put them
all in the MDC as selectable options.

It may get merged into a DMPI / Delivery Profile / notes section when such time
arises

For now, we add a free form text field below the stations and above BULLSEYE /
RAMROD information.